Removed stub from BMIWidget, replaced it with db

+++ Models/PrivateArea/Widgets/BMIWidget.go
@@ -1,6 +1,6 @@
 package Widgets
 
 type BMIWidget struct {
-	Index  float32 `json:"index"`
+	Index  float64 `json:"index"`
 	Result string  `json:"result"`
 }
